feat: 단일 코스조회 로직 추가 및 서비스 분리 [#34]
