+++ backend/src/main/java/com/mohaeng/backend/member/domain/CourseBookMark.java
@@ -18,4 +18,12 @@ public class CourseBookMark {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "course_id")
     private Course course;
+
+    public void setCourse(Course course) {
+        this.course = course;
+    }
+
+    public void setMember(Member member) {
+        this.member = member;
+    }
 }
