+++ backend/src/main/java/com/mohaeng/backend/member/service/MemberService.java
@@ -1,7 +1,9 @@
 package com.mohaeng.backend.member.service;
 
 import com.mohaeng.backend.common.BaseResponse;
+import com.mohaeng.backend.course.domain.Course;
 import com.mohaeng.backend.course.repository.CourseRepository;
+import com.mohaeng.backend.member.domain.CourseBookMark;
 import com.mohaeng.backend.member.domain.Member;
 import com.mohaeng.backend.member.dto.response.MyPageCourseBookMarkDto;
 import com.mohaeng.backend.member.repository.MemberRepository;
@@ -10,6 +12,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -21,19 +24,5 @@ public class MemberService {
     public Member findByEmail(String email) {
         return memberRepository.findByEmail(email)
                 .orElseThrow(() -> new IllegalArgumentException("INVALID_USER"));
-    }
-
-    public ResponseEntity findAllBookMarkCourse(Member member) {
-        List<MyPageCourseBookMarkDto> data = member.getCourseBookMarkList().stream()
-                .filter(m -> courseRepository.findById(m.getId()).isPresent())
-                .map(m -> new MyPageCourseBookMarkDto(m.getId(),
-                        m.getCourse().getId(), m.getCourse().getTitle(),
-                        m.getCourse().getRegion(), m.getCourse().getContent(),
-                        m.getCourse().getIsPublished(), m.getCourse().getCreatedDate(),
-                        m.getCourse().getModifiedDate()))
-                .collect(Collectors.toList());
-
-
-        return ResponseEntity.ok().body(BaseResponse.success("OK", data));
     }
 }
